pyproject.toml
[tool.poetry]
name = "pufferblow"
version = "0.0.1-beta"
description = "The official API for PufferBlow"
authors = ["ramsy0dev <ramsy0dev@gmail.com>"]
license = "GPL-3.0"
readme = "README.md"
packages = [
    {include = "pufferblow"},
    {include = "docs"}
]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.115.12"
typer = "^0.20.0"
rich = "^14.2.0"
uvicorn = "^0.38.0"
pycryptodome = "^3.23.0"
loguru = "^0.7.3"
pytz = "^2025.2"
psycopg2-binary = "^2.9.10"
pytest = "^8.4.2"
sqlalchemy = "^2.0.44"
websockets = "^15.0"
httpx = "^0.28.1"
sphinx = "^8.1.3"
sphinx-book-theme = "^1.1.2"
sphinx-autobuild = "^2024.10.3"
sphinx-favicon = "^1.0.1"
myst-parser = "^4.0.0"
sqlalchemy-utils = "^0.42.0"
orjson = "^3.11.3"
basedpyright = "^1.31.7"
pymemcache = "^4.0.0"
cryptography = "^46.0.2"
bcrypt = "^5.0.0"
watchdog = "^6.0.0"
python-magic = "^0.4.27"
pillow = "^12.0.0"
pytest-asyncio = "^1.2.0"
python-multipart = "^0.0.20"
psutil = "^7.1.0"
aiortc = "^1.14.0"
boto3 = "^1.34.0"

[tool.poetry.group.dev.dependencies]
loguru = "^0.7.3"
gunicorn = "^23.0.0"
black = "^24.10.0"
isort = "^5.13.2"
ruff = "^0.8.0"
pre-commit = "^4.0.1"

[tool.poetry.plugins."console_scripts"]
pufferblow = "pufferblow:run"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pyprojectx]
main = ["pdm==2.12.4"]  # installs pdm to your project instead of globally

[tool.pdm.dev-dependencies]  # or the poetry equivalent
dev = [
    "basedpyright", # you can pin the version here if you want, or just rely on the lockfile
]

[tool.basedpyright]
# many settings are not enabled even in strict mode, which is why basedpyright includes an "all" option
# you can then decide which rules you want to disable
typeCheckingMode = "standard"

[tool.black]
line-length = 88
target-version = ['py310']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
  | __pycache__
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["pufferblow"]
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.ruff]
target-version = "py310"
line-length = 88
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501", # line too long, handled by black
    "B008", # do not perform function calls in argument defaults
    "C901", # too complex
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]
"tests/**/*" = ["B011"] # assert false
